+++ client/web/src/insights/pages/creation/search-insight/SearchInsightCreationPage.tsx
@@ -7,11 +7,12 @@ import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { asError } from '@sourcegraph/shared/src/util/errors'
+import { useLocalStorage } from '@sourcegraph/shared/src/util/useLocalStorage'
 
 import { AuthenticatedUser } from '../../../../auth'
 import { Page } from '../../../../components/Page'
 import { PageTitle } from '../../../../components/PageTitle'
-import { FORM_ERROR } from '../../../components/form/hooks/useForm'
+import { FORM_ERROR, FormChangeEvent } from '../../../components/form/hooks/useForm'
 import { InsightsApiContext } from '../../../core/backend/api-provider'
 import { addInsightToCascadeSetting } from '../../../core/jsonc-operation'
 
@@ -20,6 +21,7 @@ import {
     SearchInsightCreationContentProps,
 } from './components/search-insight-creation-content/SearchInsightCreationContent'
 import styles from './SearchInsightCreationPage.module.scss'
+import { CreateInsightFormFields } from './types'
 import { getSanitizedSearchInsight } from './utils/insight-sanitizer'
 
 export interface SearchInsightCreationPageProps
@@ -38,6 +40,11 @@ export const SearchInsightCreationPage: React.FunctionComponent<SearchInsightCre
     const { platformContext, authenticatedUser, settingsCascade, telemetryService } = props
     const { updateSubjectSettings, getSubjectSettings } = useContext(InsightsApiContext)
     const history = useHistory()
+
+    const [initialFormValues, setInitialFormValues] = useLocalStorage<CreateInsightFormFields | undefined>(
+        'insights.search-insight-creation',
+        undefined
+    )
 
     useEffect(() => {
         telemetryService.logViewEvent('CodeInsightsSearchBasedCreationPage')
@@ -65,6 +72,9 @@ export const SearchInsightCreationPage: React.FunctionComponent<SearchInsightCre
                 await updateSubjectSettings(platformContext, subjectID, editedSettings).toPromise()
 
                 telemetryService.log('CodeInsightsSearchBasedCreationPageSubmitClick')
+
+                // Clear initial values if user successfully created search insight
+                setInitialFormValues(undefined)
                 history.push('/insights')
             } catch (error) {
                 return { [FORM_ERROR]: asError(error) }
@@ -72,13 +82,26 @@ export const SearchInsightCreationPage: React.FunctionComponent<SearchInsightCre
 
             return
         },
-        [telemetryService, history, updateSubjectSettings, getSubjectSettings, platformContext, authenticatedUser]
+        [
+            authenticatedUser,
+            getSubjectSettings,
+            updateSubjectSettings,
+            platformContext,
+            telemetryService,
+            setInitialFormValues,
+            history,
+        ]
     )
+
+    const handleChange = (event: FormChangeEvent<CreateInsightFormFields>): void => {
+        setInitialFormValues(event.values)
+    }
 
     const handleCancel = useCallback(() => {
         telemetryService.log('CodeInsightsSearchBasedCreationPageCancelClick')
+        setInitialFormValues(undefined)
         history.push('/insights')
-    }, [history, telemetryService])
+    }, [history, setInitialFormValues, telemetryService])
 
     // TODO [VK] Move this logic to high order component to simplify logic here
     if (authenticatedUser === null) {
@@ -108,9 +131,11 @@ export const SearchInsightCreationPage: React.FunctionComponent<SearchInsightCre
                 className="pb-5"
                 dataTestId="search-insight-create-page-content"
                 settings={settingsCascade.final}
+                initialValue={initialFormValues}
                 organizations={orgs}
                 onSubmit={handleSubmit}
                 onCancel={handleCancel}
+                onChange={handleChange}
             />
         </Page>
     )