+++ client/web/src/integration/graphQlResults.ts
@@ -50,6 +50,7 @@ export const commonWebGraphQlResults: Partial<WebGraphQlOperations & SharedGraph
                     },
                     settingsURL: '/site-admin/global-settings',
                     viewerCanAdminister: true,
+                    allowSiteSettingsEdits: true,
                 },
             ],
             final: JSON.stringify({}),